requirements.txt
altgraph==0.17.5
lxml==6.0.2
odfdo==3.17.6
packaging==25.0
pip-autoremove==0.10.0
psutil==7.1.3
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
setuptools==80.9.0
libvespy~=0.1.0
click~=8.3.1
vesperando_core~=0.2.0
PyYAML~=6.0.3